requirements.txt
torch
transformers
timm
Pillow
pandas
torchvision
